- Split `util_bot/error_decomp.{h,cc,inl}` out of `stabilizers/conversions.{h,cc,inl}` - Split `util_top/circuit_inverse_unitary.{h,cc,inl}` out of `stabilizers/conversions.{h,cc,inl}`
