+++ file_lists/source_files_no_main
@@ -89,3 +89,5 @@ src/stim/simulators/transform_without_feedback.cc
 src/stim/simulators/vector_simulator.cc
 src/stim/stabilizers/conversions.cc
 src/stim/stabilizers/flex_pauli_string.cc
+src/stim/util_bot/error_decomp.cc
+src/stim/util_top/circuit_inverse_unitary.cc
