+++ file_lists/test_files
@@ -81,4 +81,6 @@ src/stim/stabilizers/tableau.test.cc
 src/stim/stabilizers/tableau_iter.test.cc
 src/stim/str_util.test.cc
 src/stim/test_util.test.cc
+src/stim/util_bot/error_decomp.test.cc
+src/stim/util_top/circuit_inverse_unitary.test.cc
 src/stim_included_twice.test.cc
